pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dltr"
version = "0.1.0"
description = "Chinese scene text recognition and semantic analysis thesis project."
readme = "PLAN.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Orange" }
]
dependencies = [
  "matplotlib>=3.9.0",
  "numpy>=1.26.0",
  "opencv-python-headless>=4.10.0.84",
  "pandas>=2.2.2",
  "pillow>=10.4.0",
  "pyyaml>=6.0.2",
  "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.2",
]
demo = [
  "streamlit>=1.38.0",
]
infer-cpu = [
  "onnxruntime>=1.19.0",
]
infer-cu = [
  "onnxruntime-gpu>=1.19.0",
]
train-cu = [
  "torch>=2.5.0",
  "torchvision>=0.20.0",
]
viz = [
  "plotly>=5.24.0",
]

[project.scripts]
dltr = "dltr.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
config-settings = { editable_mode = "compat" }

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
